+++ Src/PRuntimes/PSymbolicRuntime/src/main/java/psymbolic/runtime/values/PBool.java
@@ -1,4 +1,4 @@
-package psymbolic.runtime.concretevalues;
+package psymbolic.runtime.values;
 
 public class PBool extends PValue<PBool>{
     // stores the int value
